aiproxy/requirements.txt
litellm[proxy]>=1.44.7
psycopg2-binary==2.9.9
langfuse==2.53.2
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.2
opentelemetry-instrumentation-logging==0.48b0
typing-extensions